Update solution for mean-squared-error

+++ mean-squared-error/mean-squared-error.py
@@ -5,7 +5,7 @@ def mean_squared_error(y_pred, y_true):
     Returns: float MSE
     """
     # Write code here
-    y_pred = np.array(y_pred, dtype=float)
-    y_true = np.array(y_true, dtype=float)
+    y_pred = np.array(y_pred)
+    y_true = np.array(y_true)
     
     return np.mean((y_pred - y_true)**2)
